% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtheory-package.R
\docType{data}
\name{Brennan.3.2}
\alias{Brennan.3.2}
\title{Brennan's (2001) Table 3.2}
\source{
Brennan, R. L. (2001). \emph{Generalizability theory}. New York: Springer.
}
\arguments{
\item{Brennan.3.2}{data frame in long format with a column for item scores and columns for sources of variance}
}
\description{
Brennan's (2001) table 3.2 (synthetic data set number 4) contains item scores from a person * (rater : task) generalizability (G) study.  Note that the data are in long format (one item score per row) and the facets are factors.
}
\examples{
data(Brennan.3.2)
head(Brennan.3.2)
sapply(Brennan.3.2, class)

}
\keyword{data}

