% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtheory-package.R
\docType{package}
\name{gtheory-package}
\alias{gtheory}
\alias{gtheory-package}
\title{Apply generalizability theory with R.}
\description{
\code{gtheory} provides functions for estimating variance components, generalizability coefficients, universe scores, and standard errors when observed scores contain variation from one or more measurement facets (e.g., items and raters).
}
\details{
\code{gtheory} was designed to apply univariate and multivariate generalizability (G).  Use it to conduct G and decision (D) studies with balanced or unbalanced data.  Future releases will provide the means for calculating universe scores under univariate G theory or within strata, as well as universe-score profiles (a.k.a. augmented true scores) and composite scores under multivariate G theory.  Functions to find an optimal study design and to calculate conditional error variance are also in the works.
}
\author{
Christopher T. Moore \email{moor0554@umn.edu}
}
\seealso{
\url{http://EvaluationDashboard.com}
}
\keyword{package}

