% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtheory-package.R
\docType{data}
\name{Rajaratnam.2}
\alias{Rajaratnam.2}
\title{Rajaratnam, Cronbach and Gleser's (1965) Table 2}
\source{
Rajaratnam, N., Cronbach, L. J., & Gleser, G. C. (1965). Generalizability of stratified-parallel tests. \emph{Psychometrika}, \emph{30}(1), 39-56.
}
\arguments{
\item{Rajaratnam.2}{data frame in long format with a column for item scores and columns for sources of variance}
}
\description{
Rajaratnam, Cronbach and Gleser's (1965) table 2 contains item scores from a person * (item : subscale) multivariate generalizability (G) study.  The same data are found in Brennan (2001) as table 9.1.  Note that the data are in long format (one item score per row) and the facets are factors.
}
\examples{
data(Rajaratnam.2)
head(Rajaratnam.2)
sapply(Rajaratnam.2, class)

}
\references{
Brennan, R. L. (2001). \emph{Generalizability theory}. New York: Springer.
}
\keyword{data}

