% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.components.R
\name{update.components}
\alias{update.components}
\title{update.components}
\usage{
\method{update}{components}(object, data, colname.objects, colname.scores, ...)
}
\description{
update.components
}
\keyword{internal}

