% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.gstudy.R
\name{update.gstudy}
\alias{update.gstudy}
\title{update.gstudy}
\usage{
\method{update}{gstudy}(object, data, colname.objects, colname.scores,
  colname.strata = NULL, ...)
}
\description{
update.gstudy
}
\keyword{internal}

